+++ app/views/layout.blade.php
@@ -38,7 +38,9 @@
                 <!-- Collect the nav links, forms, and other content for toggling -->
                 <div class="collapse navbar-collapse navbar-ex1-collapse">
                     <ul class="nav navbar-nav side-nav">
-                        <li class="active"><a href="{{ url() }}"><i class="fa fa-dashboard"></i> Dashboard</a></li>
+                        @if(Route::currentRouteName() == 'doolox.dashboard')<li class="active">@else<li>@endif
+                            <a href="{{ url() }}"><i class="fa fa-dashboard"></i> Dashboard</a>
+                        </li>
 @if (Auth::check())
                         <li class="dropdown">
                             <a href="javascript: void null;" class="dropdown-toggle" data-toggle="dropdown"><i class="fa fa-cog"></i> Settings <b class="caret"></b></a>
@@ -49,7 +51,9 @@
                                 <li><a href="{{ url() }}">FTP Servers</a></li>
                             </ul>
                         </li>
-                        <li><a href="{{ url() }}"><i class="fa fa-user"></i> Account</a></li>
+                        @if(Route::currentRouteName() == 'user.account')<li class="active">@else<li>@endif
+                            <a href="{{ route('user.account') }}"><i class="fa fa-user"></i> Account</a>
+                        </li>
 @endif
                     </ul>
                     <ul class="nav navbar-nav navbar-right navbar-user">