+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Doolox Changelog
 
 * First version
 * Custom filter for checking ownership
+* Account management